requirements.txt
django-crispy-forms==1.10.0
django-phone-field==1.8.1
Pillow==10.2.0
pytz==2020.4
qrcode==6.1
sqlparse==0.4.1
numpy==1.26.3
twilio==7.9.2
stripe==3.5.0
geopy==2.4.1
shortuuid==1.0.11
folium==0.15.1
pyzbar==0.1.9